# server/go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Informacje o pliku w systemie",
        "title": "ProjecTea File Service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/files": {
            "get": {
                "description": "Zwraca listę wszystkich plików przesłanych do serwera. Zawiera informacje o każdym pliku takie jak nazwa, rozmiar, data wysłania i URL do pobrania.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Lista wszystkich plików",
                "responses": {
                    "200": {
                        "description": "Lista plików",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/definitions/main.FileInfo"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Błąd serwera",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/files/download/{fileId}": {
            "get": {
                "description": "Pozwala na pobranie pliku na podstawie jego identyfikatora. Plik jest zwracany jako załącznik do pobrania.",
                "produces": [
                    "application/octet-stream"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Pobieranie pliku",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identyfikator pliku",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plik pomyślnie pobierany",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Plik nie znaleziony",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/files/upload": {
            "post": {
                "description": "Pozwala na przesłanie pliku do serwera. Maksymalny rozmiar pliku to 100MB. Plik otrzymuje unikalny identyfikator opparty na czasowniku nanosekund i nazwie oryginału.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Przesyłanie pliku",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Plik do przesłania",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plik pomyślnie przesłany",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Błąd: brak pliku lub plik za duży",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Błąd serwera",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/files/{fileId}": {
            "delete": {
                "description": "Pozwala na usunięcie pliku z serwera na podstawie jego identyfikatora. Po usunięciu plik nie będzie dostępny do pobrania.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Files"
                ],
                "summary": "Usuwanie pliku",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Identyfikator pliku",
                        "name": "fileId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plik pomyślnie usunięty",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "boolean"
                            }
                        }
                    },
                    "404": {
                        "description": "Plik nie znaleziony",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Błąd serwera",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.FileInfo": {
            "description": "Informacje o pliku w systemie",
            "type": "object",
            "properties": {
                "downloadUrl": {
                    "type": "string",
                    "example": "/api/files/download/1704067800000000000_document.pdf"
                },
                "fileId": {
                    "type": "string",
                    "example": "1704067800000000000_document.pdf"
                },
                "mimeType": {
                    "type": "string",
                    "example": "application/pdf"
                },
                "name": {
                    "type": "string",
                    "example": "document.pdf"
                },
                "size": {
                    "type": "integer",
                    "example": 1024000
                },
                "uploadAt": {
                    "type": "string",
                    "example": "2026-04-29T10:30:00Z"
                }
            }
        }
    }
}